+++ examples/with-next-mint-nft/CHANGELOG.md
@@ -1,5 +1,19 @@
 # with-next-mint-nft
 
+## 0.0.32
+
+### Patch Changes
+
+- Updated dependencies [a1d6776]
+  - @rainbow-me/rainbowkit@0.8.1
+
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies [6b37050]
+  - @rainbow-me/rainbowkit@0.8.0
+
 ## 0.0.30
 
 ### Patch Changes